feat(*): adds support for including related resources
Closes #1
