+++ mocha.opts
@@ -1,2 +1,3 @@
 --require ts-node/register
+--check-leaks
 spec/**/*.spec.ts
